% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLUcolors.R
\name{SLU_cols}
\alias{SLU_cols}
\title{Function to extract SLU colors as hex codes}
\usage{
SLU_cols(...)
}
\arguments{
\item{...}{Character names of SLU_colors}
}
\description{
Function to extract SLU colors as hex codes
}
