% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLUcolors.R
\name{SLU_pal}
\alias{SLU_pal}
\title{Return function to interpolate a SLU color palette}
\usage{
SLU_pal(palette = "all", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in SLU_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a SLU color palette
}
