% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLUcolors.R
\name{SLUpalette}
\alias{SLUpalette}
\title{Work with SLU colors}
\usage{
SLUpalette(palette, color_names = TRUE)
}
\arguments{
\item{palette}{character or numeric defining the desired color palette. Allowed values  "all", "red",
"green", "blue", "yellow", "grey" and "wong_234516" or a number 1:5.}

\item{color_names}{logical return named vector or not. Default TRUE}
}
\value{
a named vector of color specifications.
}
\description{
Get different variants of color palettes containing official SLU colors. Also avaliable is the "wong_234516"
palette with non SLU colors but adapted to colorblindness. "wong_234516" will be deprecated if/when a similar
and official palette becomes available.
}
\details{
\code{SLUpalette(palette)} returns different variants of the official SLU color palette.
}
\examples{
opal <- palette(SLUpalette("red")) # set default palette to SLU "red" and save original palette
plot(Sepal.Width ~ Sepal.Length, data = iris, pch = 16, col = Species)
palette(SLUpalette("blue")) # set default palette to SLU "blue"
plot(Sepal.Width ~ Sepal.Length, data = iris, pch = 16, col = Species)
palette(opal)

}
