% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_SLU.R
\name{theme_SLU}
\alias{theme_SLU}
\title{SLU theme for ggplot2}
\usage{
theme_SLU(palette, dark = TRUE)
}
\arguments{
\item{palette}{character or numeric defining the desired color palette}

\item{dark}{logical return theme for dark or light color scheme. Default dark.}
}
\value{
a ggplot theme using SLU colors
}
\description{
SLU theme for ggplot2
}
\examples{
library(ggplot2)
p1 <- ggplot(iris, mapping = aes(x = Sepal.Width, y = Sepal.Length))

p1 + labs(title = "Grey on green") +
  geom_point(color = SLUpalette("grey")["Glimmer"]) +
  theme_SLU("green")
}
