% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLUcolors.R
\name{SLUpalette}
\alias{SLUpalette}
\title{Get different variants of color palettes containing official SLU colors}
\usage{
SLUpalette(palette)
}
\arguments{
\item{patette}{character or numeric defining the desired color palette}
}
\value{
a named vector of color specifications.
}
\description{
\code{SLUpalette(palette)} returns different variants of the official SLU color palette.
}
\examples{
opal <- palette(SLUpalette("red")) # set default palette to SLU "red" and save original palette
plot(Sepal.Width ~ Sepal.Length, data = iris, pch = 16, col = Species)
palette(SLUpalette("blue")) # set default palette to SLU "blue"
plot(Sepal.Width ~ Sepal.Length, data = iris, pch = 16, col = Species)
palette(opal)

}
