% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_SLU.R
\name{theme_SLU}
\alias{theme_SLU}
\title{SLU theme for ggplot2}
\usage{
theme_SLU(palette, dark = TRUE)
}
\arguments{
\item{palette}{character or numeric defining the desired color palette}

\item{dark}{logical return theme for dark or light color scheme. Default dark.}
}
\value{
a ggplot theme using SLU colors
}
\description{
SLU theme for ggplot2
}
